GEODE-6112: Gfsh should respond to SIGINT
Gfsh now responds to SIGINT by cancelling the foreground process (the
standard response to SIGINT), except when the user has entered text in
the command prompt. If text is entered, the entered text is ignored and
a new prompt is returned. The entered text will be deleted from the
history.