GEODE-6121 Port SnappyData DirectChannel comms changes to Geode
Not ready for review!

Initial port of SnappyData peer-to-peer communication changes to Geode.
See https://github.com/SnappyDataInc/snappy-store/commit/49222766a35c99f4853b7b6661457f37b3c39349

The current changes have connection-streaming and conserve-sockets=false
as the default for testing but I will be changing this.  The connection-streaming
feature will be an opt-in thing at first and probably be marked experimental.