GEODE-2113 Implement SSL over NIO
This removes old-I/O use in TCPConduit peer-to-peer communications.
This was used for SSL/TLS secure commuications but Java has had an
SSLEngine implementation that allows you to implement secure
communications on new-I/O SocketChannels or any other transport
mechanism.

A new NioSSLEngine class wraps the JDK's SSLEngine and provides the SSL
handshake as well as encryption/decryption of messages.  SocketCreator
performs the SSL handshake and returns a NioSslEngine that TCPConduit
then uses for messaging.

I've also done a lot of cleanup of compilation warnings in
Connection.java and removed references to "NIO".  The primary SSL/TLS
changes in that class are in writeFully (renamed from nioWriteFully)
and processBuffer (renamed from processNIOBuffer).

Porting client/server to use NioSSLEngine will be done under a separate
ticket and a different version of NioEngine may be created to secure
UDP messaging.