GEODE-6112: Gfsh should respond to SIGINT (#2964)
Gfsh now responds to SIGINT by clearing the prompt. It does not print a
new prompt, as that implementation caused issue with the Gfsh History.
This is the behavior regardless of if there is text entered in the
prompt.

When a command is running, the command can be sent to the background
using SIGINT. For example, when running the start server command, if
the GfshSignalHandler receives a SIGINT, the command will be interrupted
but the server start will continue in the background.