GEODE-6215: Making session metadata region distributed
The meta data region used for session replication was converted to
Scope.LOCAL by mistake in 7ca571c5. This broke the logic to
automatically create the session region on all peers when starting up
the session replication service.

Changing the scope back to DISTRIBUTED_ACK for this metadata region, and
adding tests that session region creation works.

Co-Authored-By: Dan Smith <dsmith@pivotal.io>