GEODE-6236 fix inconsistent handling of versionNumber
In a few build scripts versionNumber was obtained by parsing gradle.properties directly. This approach gave incorrect value if that property had been overridden with a -P on the gradle command line.

Now the overrides are considered before the properties file, mimicking the value gradle would choose in all cases.