leave the member unsuspected after final check fails
If the final check fails and we're not going to remove the suspect from
the distributed system we need to leave it in an "unsuspected" state
locally so that the background monitoring thread will look at it again.

Also, if the final check failed in the membership coordinator there's no
point in doing another check so we move directly to removing the
suspect.