GEODE-2113 Implement SSL over NIO
This commit fixes problems encountered in the previous implementation:

- position tracking for MsgReader has been moved into NioSslEngine and
NioPlainEngine because the Ssl engine sometimes needs to allocate a new
buffer and adjust the tracked read/process positions in the buffer

- the same buffer expansion methods in Buffers was being used for
read-side and write-side buffers but buffers being used for reads need
to be handled differently than buffers being used for writing

The first commit is the original implementation.  The second commit
fixes the problems in that commit.  If you've already reviewed the old
code you can probably just look at the second commit.

There is commented-out debugging code that I'll remove when these
changes pass the Pull Request Stress Tests.