GEODE-6297: Add Process info to Launcher status failure msg
This will provide some details about if the process is actually
alive or not when FileProcessController.status throws
IllegalStateException.

When the test fails again, if Process is not alive then the
underlying cause is probably something like a BindException
preventing the startup of the Locator process.

This info will also apply to GEODE-6183 which is nearly
identical.

I'm including the change in ServerLauncherRemoteIntegrationTestCase
as well in case there's ever a similar failure starting up a
server.