GEODE-6244 Healthy member kicked out by Sick member when final-check fails
The initial fix caused a problem that prevented election of a new
membership coordinator in a certain case.  The case was a view
with nodes [A, B, C, D, E] where C was the coordinator.  Node A had
crashed and the crash had been detected by B.  Node C then left the
cluster, sending a Leave message to B.  B's JoinLeave did not know about
the HealthMonitor's decision that A was crashed and did not become the
new coordinator.

This commit makes B's JoinLeave pay attention to the crashed-member set
in the HealthMonitor when deciding whether to become the membership
coordinator for the cluster.