GEODE-6329: Simplify logic for reporting present disk stores
The logic for reporting what disk stores were actually running on a
node, for the purposes of not reporting those disk stores as missing,
was overly complicated. Due to that, it only included disk stores that
actually had any persistent regions/buckets on that node, rather than
all disk stores.

Refactoring the code to just list all disk stores.