GEODE-6329: Make sure member is removed from persistent view
When we start up a member that has persistent data that it should delete
(due to already established redundancy), the member was not correctly
removing members from it's persistent view in
PersistenceAdvisorImpl.updateMembershipView. This was due to the fact
that the destroyOfflineData method was clearing the recoveredMembers
set, but not the underlying persistentMemberView.

The persistentMemberView probably should be preserved, in case the
member is killed and restarted before it gets a new view. So we should
reset the recoveredMembers set to the correct value.