GEODE-6329, GEODE-6328: Fix issues with show missing-disk-stores (#3136)
* Using consistent logic for finding missing disk stores

The gfsh command for finding missing disk stores was not ignoring
present disk stores. However, the JMX command was. The gfsh command was also not deduplicating the missing disk stores. Changing the gfsh command to use the JMX logic.

* Making sure member is removed from persistent view

When we start up a member that has persistent data that it should delete
(due to already established redundancy), the member was not correctly
removing members from it's persistent view in
PersistenceAdvisorImpl.updateMembershipView. This was due to the fact
that the destroyOfflineData method was clearing the recoveredMembers
set, but not the underlying persistentMemberView.

The persistentMemberView probably should be preserved, in case the
member is killed and restarted before it gets a new view. So we should
reset the recoveredMembers set to the correct value.

* Simplifying the logic for reporting present disk stores

The logic for reporting what disk stores were actually running on a
node, for the purposes of not reporting those disk stores as missing,
was overly complicated. Due to that, it only included disk stores that
actually had any persistent regions/buckets on that node, rather than
all disk stores.

Refactoring the code to just list all disk stores.