GEODE-6363 LocatorUDPSecurityDUnitTest
Changes in GMSJoinLeave prevent a restarting locator from becoming
membership coordinator if its findCoordinatorFromView method indicates
that it received responses from processes that are members of a
distributed system.  It now also waits for more than one response to its
requests to find the coordinator from other members so that it gets a
better picture of who's out there that might be the current coordinator.