GEODE-6323 Fix dependency tracking for manifest jars (#3126)
Because `doFirst` registers an action to be executed, it is not valid to be used as part of a task's configuration, causing aberrant behavior when combined with Gradle's output caching.

If "configuration" must occur at execution time, then that configuration needs to be the output of a task on which the "configured" task depends.  We have done that here to correctly generate the classpath in our dependency jars.  For more information, see
https://guides.gradle.org/using-build-cache/#suggestions_for_authoring_your_build

Co-authored-by: Robert Houghton <rhoughton@pivotal.io>
Co-authored-by: Patrick Rhomberg <prhomberg@pivotal.io>