GEODE-6380: Making gateway stat id fields immutable
The stat ids in GatewaySenderStats were being reset in the static
intitializer for AsyncEventQueueStats. This likely resulted in all of
the stats that are listed in GatewaySenderStats after BATCHES_RESIZED
being recorded against the wrong stat name, because that stat was not
part of the AsyncEventQueueStats, and so the overriden were incorrect.