GEODE-6389 CI Failure: ConcurrentWANPropagation_1_DUnitTest.testReplicatedSerialPropagation
I found a place where we were releasing pooled buffers twice.  The place
to release the buffer is in asyncClose but we were also doing it just
after calling that method in the run() method if the connection wasn't
thread-owned.  Both GEODE-6389 and GEODE-6377 were using
conserve-sockets=true so I think this is the source of the problem.

The GEODE-6389 failure, in particular, could only have come from
improper use of the buffer pool because it occured in a MessageStreamer
and they don't involve multithreaded use of a buffer.