GEODE-6389 CI Failure: ConcurrentWANPropagation_1_DUnitTest.testReplicatedSerialPropagation_withoutRemoteSite
This removes the sharing of buffers between threads in Connection.java
and MsgReader.java.  The Buffers buffer pooling mechanism isn't up to
handling sharing of buffers between threads and I determined that it
wasn't really necessary to retain the handshake buffer for follow-up TLS
communications.

ClusterCommunicationsDUnitTest already covers testing of this change.

This closes #3209