GEODE-6422: Fix test timeout issue
This test was timing out when run in CI. Since it never fails locally,
the issue appears to be with resource contention. Increasing the timeout
should reduce the frequency of failures as it gives the test more time
to spin up the required threads. This test will still fail if it fails
to start two threads within a minute.