GEODE-5013 Replace org.json with Jackson in GfJsonObject
This removes org.json as a dependency in geode-core and associated test
modules.  GfJsonObject and GfJsonArray are ported to use the same basic
Jackson ObjectMapper setup as QueryResultFormatter so that we now have a
uniform mechanism for managing JSON documents for querying/gfsh.

The geode-web-api and geode-pulse modules still use the geode-json
module.  When/if they are ported to use GfJsonObject/Array we can delete
the geode-json module.