GEODE-6447: Honor the PARALLEL_GRADLE flag in builds
The concourse jobs where we tried to turn of parallel gradle were still
running gradle in parallel, because it's the default. Pass --no-parallel
if the job is configured not to run in parallel.

This was causing port/file related conflicts in the windows tests, which
are set to not run in parallel.