GEODE-6271: Restoring pdx type copying behavior in client
Some users were relying on having PDX types copied from one cluster to
another in the client. This logic was removed (accidentally?) in
cf0b378429. Restoring the original behavior to copy a type from one
cluster to another when it is created.

Note that receiving a type through a read still does not result in
copying a type from one cluster to another, so this topology is not
completely supported even with this fix.