GEODE-6526: Removing call to removeTombstone during entry destroy
During destroy entry (tombstone) under region entry lock, if the
entry had a lower region version than the recorded gc version for
that member, the entry was removed immediately which could result
in dead-lock with tombstone gc thread.

Instead of removing the entry during destroy, it was scheduled to
remove during tombstone gc.