GEODE-6557: Moving geode-web classes to geode-web
Moving the classes in geode-core that really should be in geode-web.
This removes optional dependencies from geode-core and puts these
classes where their tests are.

As part of this change, SerializableObjectHttpMessageConverter is now
duplicated in geode-web as ServerSerializableObjectHttpMessageConverter.
This class was previously put in both the geode-web.war file and the
gfsh-dependencies by creating a separate jar with all of the geode-web
classes as part of the geode-core build.

This class is used by both gfsh and geode-web. But geode-web needs its
own copy so that the class is loaded by the geode-web war classloader in
order to ensure the class is linked with the spring-web classes coming
from the geode-web classloader.