GEODE-6587: fix race in useFirst
useFirst had a race condition in which the connection it
returned might not match the predicate. This is because
the predicate was only checked while the connection will
still not active. Now a recheck is done after activating
the connection. If, after activation it does not match, the
connection is added back on the end of the deque.