GEODE-6646 - CI failure in serverRestarsAfterLocatorReconnects
In this test there is a locator and there are two servers.  The
locator and second servers are forced out of the cluster and
auto-reconnect.  In the failure the first server becomes membership
coordinator but the other processes fail to join its cluster when
they restart.  They are misconfigured to have a max-wait-time-reconnect
that is too short (5 seconds instead of 60) to allow the first server to detect failures
before the failed server and locator start trying to reconnect.

The fix is in InternalDistributedSystem and ensures that the
waiting period before attempting to reconnect is a sufficient
multiple of the member-timeout setting.