GEODE-6387: Extract MemberLevelStatsTest
MemberLevelStatsIntegrationTest setUp creates a cache, but the majority
of the tests do not even use or require that cache. So, it's a good
idea to extract those to the new unit test MemberLevelStatsTest.

These changes further isolate the testing in MemberLevelStatsTest to
avoid any potential flakiness.

We also clean up MemberLevelStatsIntegrationTest to prevent potential
causes of flakiness, but ultimately GEODE-6387 is now fixed because
only one test remains so it should be impossible for another test to
leave around an incompatible distributed system.