GEODE-6904: Fix hang on locator when restarting due to lock
Only hold the locatorLock while reading the locator variable in
restartWithDS. Holding this lock while recovering cluster configuration
was causing other messages to the locator, including the StartupMessage,
to block waiting for this lock.

In addition, waiting forever for StartupMessage responses.
StartupOperation had a 15 second timeout, after which it would stop
waiting. That resulted in not having correct metadata about other
members in the system.

Co-Authored-By: Ernest Burghardt <eburghardt@pivotal.io>