implemented backward-compatibility in serialization
The Geode MembershipView is no longer serializable.  The only view that
is serializable is now GMSMembershipView, which has the same on-wire
form as the old MembershipView class.

GMSMember now has the same on-wire format as InternalDistributedMember.
Prior to this branch GMSMember was never serialized except with its
writeEssentialData method, which is used by the Messenger service.