GEODE-7012: Execute the StartupMessage in the waiting pool (#3844)
Until we get a response to a StartupMessage, we end up blocking up high
priority thread pool threads. So the StartupMessage should be processed
in a different, unbounded threadpool to avoid a distributed deadlock.

Also, added unit tests for StartupMessage and StartupResponseMessage excercising getProcessorType() to guard against future unintended alteration.