GEODE-7031 Attempts to send messages to alert listeners delays network partition detection
Decrease the socket-formation timeout for Alert listeners.  Generally
we'll already have a connection to an alert listener so the decreased
timeout won't be used.  In times where there are network problems,
though, we often have to create a new tcp/ip connection to send an alert
and we don't want these to stall for too long.