GEODE-7038: After auto-reconnect a server's multicat communications aren't working correctly
Ensure that a JoinResponseMessage is sent if multicast is enabled.  This
allows JGroupsMessenger to piggy-back a multicast message digest on the
response that the new process can install in its JGroups stack to ensure
that multicast messaging is properly initialized.

I've also replaced complex checks for whether UDP security is enabled
with a simpler check on ServiceConfig.  When UDP security is enabled we
are already sending a JoinResponseMessage and so we don't need to send
another one if multicast is enabled.