GEODE-7085: Removing the use of setVersion from RegionVersionVectorTest
This test was using RegionVersionHolder.setVersion in way that seemed to
imply the test was trying to record all of the version up to that
version.

setVersion is actually an internal method that called by itself leaves
the RegionVersionHolder in an invalid state. The tests were previously
passing, but some of the internal state of the RegionVersionHolder was
actually corrupt - bitSetVersion was negative, for example.

Instead of using setVersion, changing these tests to use recordVersion,
which is the public method used by other parts of the code.