GEODE-7072 CI Failure: WANRollingUpgradeEventProcessingMixedSiteOneCurrentSiteTwo
The DataSerializableFixedID class FinalCheckPassedMessage (with ID -158)
wasn't added until version 1.4.  Sending this message to a member of the
cluster running v1.3 will result in a deserialization error on that
member.  The error is benign but it does expose rolling-upgrade tests to
failure if suspect-string processing sees the deserialization error in
unit test output.