GEODCE-7123: Add create disk-store command option, stage-configuration
The new option, --stage-configuration, to the create disk-store command
provides the capability to define a disk-store in the cluster configuration
prior to any server members have joined the cluster. Using this option
will only update the cluster-configuration, disk-stores will not be
created on any current server members.

Servers joining the cluster (or restarting) will then inherit the new
disk-store configuration and create the appropriate directory and
disk-store files.

This is useful for persisting the PDX types to a disk-store other than
DEFAULT.

Authored-by: Ken Howe <khowe@pivotal.io>